Workaround for "no systray found" problem (issue 362)
Well, it seems there is some problem in Qt framework that prevents
system tray to be detected. Currently, (Qt 5.3.0-beta) it is not
resolved. So, introduce "trayless" mode for such a cases when Qt unable
to access the systray feature. In this mode SettingsWindow is visible
after start and when user closes the window, just terminate the
application.